rename model and workflow to models and workflows for consistency
