+++ hier_config/platforms/cisco_nxos/driver.py
@@ -1,4 +1,6 @@
-from hier_config.model import (
+from pydantic import Field
+
+from hier_config.models import (
     IdempotentCommandsAvoidRule,
     IdempotentCommandsRule,
     MatchRule,
