+++ hier_config/platforms/cisco_nxos/view.py
@@ -4,7 +4,7 @@ from re import sub
 from typing import Optional
 
 from hier_config.child import HConfigChild
-from hier_config.platforms.model import (
+from hier_config.platforms.models import (
     InterfaceDot1qMode,
     InterfaceDuplex,
     NACHostMode,
