+++ hier_config/platforms/hp_comware5/driver.py
@@ -1,4 +1,4 @@
-from hier_config.model import Platform
+from hier_config.models import Platform
 from hier_config.platforms.driver_base import HConfigDriverBase
