+++ hier_config/platforms/hp_procurve/driver.py
@@ -3,7 +3,7 @@ from collections.abc import Callable, Iterable
 from typing import Optional
 
 from hier_config.child import HConfigChild
-from hier_config.model import (
+from hier_config.models import (
     IdempotentCommandsRule,
     MatchRule,
     NegationDefaultWithRule,
