+++ hier_config/platforms/juniper_junos/driver.py
@@ -1,5 +1,5 @@
 from hier_config.child import HConfigChild
-from hier_config.model import Platform
+from hier_config.models import Platform
 from hier_config.platforms.driver_base import HConfigDriverBase
