+++ hier_config/platforms/view_base.py
@@ -4,7 +4,7 @@ from ipaddress import IPv4Address, IPv4Interface
 from typing import Optional
 
 from hier_config.child import HConfigChild
-from hier_config.platforms.model import (
+from hier_config.platforms.models import (
     InterfaceDot1qMode,
     InterfaceDuplex,
     NACHostMode,
