+++ tests/config_view/test_interface.py
@@ -1,5 +1,5 @@
 from hier_config import get_hconfig, get_hconfig_view
-from hier_config.model import Platform
+from hier_config.models import Platform
 from hier_config.platforms.hp_procurve.functions import hp_procurve_expand_range
