+++ tests/test_various.py
@@ -1,5 +1,5 @@
 from hier_config import get_hconfig, get_hconfig_driver
-from hier_config.model import Platform
+from hier_config.models import Platform
 
 
 def test_issue104() -> None:
